refactor(admin): split admin panel into modular feature components
- Extract tabs, modals, and shared UI into components/admin/
- Reduce pages/AdminPage.jsx to a thin access-guarded entry
- Deduplicate styles, api helper, and formatters into adminShared
