+++ client/src/components/admin/AdminPanel.jsx
@@ -1,0 +1,140 @@
+import { useCallback, useEffect, useRef, useState } from "react";
+import {
+  ArrowLeft,
+  ArrowLeftFromLine,
+  ArrowRightFromLine,
+  Gauge,
+  Moon,
+  MessageCircleMore,
+  Refresh,
+  ScrollText,
+  ShieldCog,
+  Sun,
+  Users,
+  Wrench,
+} from "../../icons/lucide.js";
+import { api } from "./adminShared.js";
+import DashboardTab from "./DashboardTab.jsx";
+import UsersTab from "./UsersTab.jsx";
+import ChatsTab from "./ChatsTab.jsx";
+import ActionsTab from "./ActionsTab.jsx";
+import LogsTab from "./LogsTab.jsx";
+
+const TABS = [
+  { id: "dashboard", label: "Dashboard", icon: Gauge },
+  { id: "users",     label: "Users",     icon: Users },
+  { id: "chats",     label: "Chats",     icon: MessageCircleMore },
+  { id: "actions",   label: "Actions",   icon: Wrench },
+  { id: "logs",      label: "Logs",      icon: ScrollText },
+];
+
+export default function AdminPanel({ user, onBack, isDark, toggleTheme }) {
+  const [tab, setTab]                 = useState("dashboard");
+  const [stats, setStats]             = useState(null);
+  const [sidebarOpen, setSidebarOpen] = useState(true);
+  const [themeAnim, setThemeAnim]     = useState(false);
+  const themeAnimRef = useRef(null);
+
+  const refreshStats = useCallback(async () => {
+    try { const d = await api.get("/api/admin/stats"); setStats(d); } catch {}
+  }, []);
+
+  useEffect(() => { refreshStats(); }, [refreshStats]);
+
+  const handleToggleTheme = () => {
+    setThemeAnim(true);
+    clearTimeout(themeAnimRef.current);
+    if (toggleTheme) toggleTheme();
+    themeAnimRef.current = setTimeout(() => setThemeAnim(false), 520);
+  };
+  useEffect(() => () => clearTimeout(themeAnimRef.current), []);
+
+  const activeTab = TABS.find((t) => t.id === tab);
+  const ActiveIcon = activeTab?.icon ?? Gauge;
+
+  return (
+    <div className="flex h-full w-full overflow-hidden bg-slate-50 dark:bg-slate-900">
+      {/* Mobile backdrop */}
+      {sidebarOpen && (
+        <div className="fixed inset-0 z-20 bg-black/30 md:hidden" onClick={() => setSidebarOpen(false)} />
+      )}
+
+      {/* Sidebar */}
+      <nav className={`
+        absolute inset-y-0 left-0 z-30 flex flex-col
+        border-r border-slate-200/80 bg-white/95 backdrop-blur-sm
+        transition-all duration-200
+        dark:border-white/5 dark:bg-slate-900/95
+        md:relative md:z-auto md:translate-x-0
+        ${sidebarOpen ? "w-56 translate-x-0" : "w-0 -translate-x-full md:w-14 md:translate-x-0"}
+      `}>
+        <div className={`flex h-12 shrink-0 items-center border-b border-slate-100 dark:border-white/5 ${sidebarOpen ? "justify-between px-3" : "justify-center"}`}>
+          {sidebarOpen && (
+            <div className="flex items-center gap-2 overflow-hidden">
+              <ShieldCog size={14} className="shrink-0 text-emerald-500" />
+              <span className="truncate text-sm font-bold text-slate-700 dark:text-slate-200">Admin Panel</span>
+            </div>
+          )}
+          <button type="button" onClick={() => setSidebarOpen((o) => !o)} title={sidebarOpen ? "Collapse" : "Expand"}
+            className="inline-flex h-8 w-8 shrink-0 items-center justify-center rounded-xl border border-transparent text-slate-400 transition hover:border-emerald-200/60 hover:bg-emerald-50/50 hover:text-emerald-600 dark:text-slate-500 dark:hover:border-emerald-500/20 dark:hover:bg-emerald-500/5 dark:hover:text-emerald-400">
+            {sidebarOpen ? <ArrowLeftFromLine size={15} /> : <ArrowRightFromLine size={15} />}
+          </button>
+        </div>
+
+        <div className="flex min-h-0 flex-1 flex-col gap-0.5 overflow-hidden p-2">
+          {TABS.map(({ id, label, icon: Icon }) => (
+            <button key={id} type="button"
+              onClick={() => { setTab(id); if (window.innerWidth < 768) setSidebarOpen(false); }}
+              title={!sidebarOpen ? label : undefined}
+              className={`flex h-9 w-full items-center rounded-xl border transition
+                ${sidebarOpen ? "gap-2.5 px-3 text-sm font-medium" : "justify-center"}
+                ${tab === id
+                  ? "border-emerald-200 bg-emerald-50 text-emerald-700 shadow-[0_0_14px_rgba(16,185,129,0.12)] dark:border-emerald-500/30 dark:bg-emerald-500/10 dark:text-emerald-300"
+                  : "border-transparent text-slate-500 hover:border-emerald-200/60 hover:bg-emerald-50/50 hover:text-emerald-700 dark:text-slate-400 dark:hover:border-emerald-500/20 dark:hover:bg-emerald-500/5 dark:hover:text-emerald-300"
+                }`}>
+              <Icon size={15} className="shrink-0" />
+              {sidebarOpen && <span className="truncate">{label}</span>}
+            </button>
+          ))}
+        </div>
+
+        <div className="shrink-0 border-t border-slate-100 p-2 dark:border-white/5">
+          <button type="button" onClick={onBack} title={!sidebarOpen ? "Exit" : undefined}
+            className={`flex h-9 w-full items-center rounded-xl border border-transparent text-rose-500 transition
+              hover:border-rose-200 hover:bg-rose-50 hover:shadow-[0_0_14px_rgba(244,63,94,0.12)]
+              dark:text-rose-400 dark:hover:border-rose-500/30 dark:hover:bg-rose-500/10
+              ${sidebarOpen ? "gap-2 px-3 text-sm font-medium" : "justify-center"}`}>
+            <ArrowLeft size={15} className="shrink-0" />
+            {sidebarOpen && <span className="truncate">Exit</span>}
+          </button>
+        </div>
+      </nav>
+
+      {/* Main content */}
+      <div className="flex min-w-0 flex-1 flex-col overflow-hidden">
+        <div className="flex h-12 shrink-0 items-center gap-2 border-b border-slate-200/80 bg-white/80 px-3 backdrop-blur-sm dark:border-white/5 dark:bg-slate-900/80">
+          <ActiveIcon size={15} className="shrink-0 text-emerald-500" />
+          <h1 className="flex-1 truncate text-sm font-semibold text-slate-700 dark:text-slate-200">{activeTab?.label}</h1>
+          <button type="button" onClick={handleToggleTheme} title={isDark ? "Switch to light mode" : "Switch to dark mode"}
+            className="inline-flex h-8 w-8 shrink-0 items-center justify-center rounded-full border border-transparent text-slate-500 transition hover:border-emerald-300 hover:bg-emerald-100 hover:text-emerald-700 dark:text-slate-400 dark:hover:border-emerald-500/30 dark:hover:bg-emerald-500/10 dark:hover:text-emerald-300">
+            {isDark
+              ? <Sun size={15} className={`icon-anim-spin-dir ${themeAnim ? "icon-theme-enter-sun" : ""}`} />
+              : <Moon size={15} className={`icon-anim-spin-left ${themeAnim ? "icon-theme-enter-moon" : ""}`} />}
+          </button>
+          <button type="button" onClick={refreshStats} title="Refresh"
+            className="inline-flex h-8 w-8 shrink-0 items-center justify-center rounded-full border border-transparent text-slate-500 transition hover:border-emerald-300 hover:bg-emerald-100 hover:text-emerald-700 dark:text-slate-400 dark:hover:border-emerald-500/30 dark:hover:bg-emerald-500/10 dark:hover:text-emerald-300">
+            <Refresh size={14} />
+          </button>
+        </div>
+
+        <div className="app-scroll min-h-0 flex-1 overflow-y-auto p-4 md:p-5">
+          {tab === "dashboard" && <DashboardTab stats={stats} onStatsChange={refreshStats} />}
+          {tab === "users"     && <UsersTab currentUser={user} onStatsChange={refreshStats} />}
+          {tab === "chats"     && <ChatsTab onStatsChange={refreshStats} />}
+          {tab === "actions"   && <ActionsTab />}
+          {tab === "logs"      && <LogsTab />}
+        </div>
+      </div>
+    </div>
+  );
+}
